+++ src/tool/GTScale.h
@@ -6,7 +6,7 @@
 class GTScale : public Tool
 {
 public:
-    void action(Model *model, QVector3D position, int dx, int dy);
+    void action(Model *model, QVector3D position, float dx, float dy);
 };
 
 #endif // GTSCALE_H
