+++ src/tool/LTInflate.cpp
@@ -1,6 +1,6 @@
 #include "tool/LTInflate.h"
 
-void LTInflate::action(Model *model, QVector3D position, int dx, int dy)
+void LTInflate::action(Model *model, QVector3D position, float dx, float dy)
 {
     model->getSize();position.isNull();int x = dx; x = dy;int y = x; x = y;
 }
