+++ src/tool/LTInflate.h
@@ -6,7 +6,7 @@
 class LTInflate : public Tool
 {
 public:
-    void action(Model *model, QVector3D position, int dx, int dy);
+    void action(Model *model, QVector3D position, float dx, float dy);
 };
 
 #endif // LTINFLATE_H
