+++ src/tool/LTMove.h
@@ -6,7 +6,7 @@
 class LTMove : public Tool
 {
 public:
-    void action(Model *model, QVector3D position, int dx, int dy);
+    void action(Model *model, QVector3D position, float dx, float dy);
 };
 
 #endif // LTMOVE_H
