+++ src/tool/LTPinch.cpp
@@ -1,6 +1,6 @@
 #include "tool/LTPinch.h"
 
-void LTPinch::action(Model *model, QVector3D position, int dx, int dy)
+void LTPinch::action(Model *model, QVector3D position, float dx, float dy)
 {
     model->getSize();position.isNull();int x = dx; x = dy;int y = x; x = y;
 }
