+++ src/tool/LTSmooth.cpp
@@ -1,6 +1,6 @@
 #include "tool/LTSmooth.h"
 
-void LTSmooth::action(Model *model, QVector3D position, int dx, int dy)
+void LTSmooth::action(Model *model, QVector3D position, float dx, float dy)
 {
     model->getSize();position.isNull();int x = dx; x = dy;int y = x; x = y;
 }
